+++ userdata/tasks.txt
@@ -1,3 +1,4 @@
 T | 0 | stuff
 D | 0 | stuff | 2024-04-04 12:00
 E | 0 | stuff | 2024-04-04 12:00 | 2024-02-02 12:30
+T | 0 | meh
